Renamed LayerType files
