+++ js/layer_type/circle_layer_type.js
