+++ js/layer_type/index.js
@@ -1,8 +1,8 @@
 'use strict';
 
 module.exports = {
-    circle: require('./circle'),
-    fill: require('./fill'),
-    line: require('./line'),
-    symbol: require('./symbol')
+    circle: require('./circle_layer_type'),
+    fill: require('./fill_layer_type'),
+    line: require('./line_layer_type'),
+    symbol: require('./symbol_layer_type')
 };
