+++ js/layer_type/symbol_layer_type.js
